+++ Types and data structures/Lab2/funcs.h
@@ -3,6 +3,6 @@
 #include <stdio.h>
 #include "types.h"
 
-void printBitmap  (MazeMatrix_t maze, uint16_t w, uint16_t h);
+void printBitmap  (MazeMatrix_t maze, int w, int h);
 
 #endif // FUNCS_H
